# swagger_api.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: DPC API
  description: Duke Programming Contest API
host: localhost
schemes:
  - http
consumes:
  - application/json
basePath: /api
produces:
  - application/json
paths:
...
  /contests/{contest}:
    get:
      summary: Retrieve contest
      parameters:
        - name: contest
          in: path
          description: Contest contestId
          required: true
          type: string
      responses:
        200:
          description: Successfully retrieved
          schema:
            $ref: '#/definitions/Contest'
  /contests/join/{contest}:
    get:
      summary: Join contest
      parameters:
        - name: contest
          in: path
          description: Contest contestId
          required: true
          type: string
      responses:
        200:
          description: Successfully joined contest
definitions:
...
  Contest:
    properties:
      contestId:
        type: string
        description: Unique contestId
      name:
        type: string
        description: Contest official name
      startDate:
        type: string
        format: dateTime
        description: Start date and time
      duration:
        type: number
        description: Duration in minutes